requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pdfplumber>=0.10.0
python-docx>=1.1.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.4.0
xgboost>=2.0.0
joblib>=1.3.0
google-generativeai>=0.5.0
